svg/struct/scripted/svg-getIntersectionList-001.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:html="http://www.w3.org/1999/xhtml"><html:link href="https://svgwg.org/svg2-draft/struct.html#__svg__SVGSVGElement__getIntersectionList" rel="help"/><html:script src="/resources/testharness.js"/><html:script src="/resources/testharnessreport.js"/><html:link href="/fonts/ahem.css" rel="stylesheet"/><style>text{font-family:Ahem;font-size:100px}</style><text x="0" y="80">Text0</text><text x="0" y="80" style="display:none">Text1</text><text x="0" y="80" pointer-events="none">Text2</text><text x="0" y="80">Text3</text><script>test(t =&gt; { const svg = document.documentElement; const rect = svg.createSVGRect(); rect.x = 10; rect.y = 70; rect.width = 20; rect.height = 20; const texts = document.querySelectorAll(&apos;text&apos;); const list = svg.getIntersectionList(rect, null); assert_array_equals(list, [texts[0], texts[3]]); }, &apos;SVGSVGElement.getIntersectionList()&apos;);</script></svg>
